fixup! KTOR-9542 Fix Netty requests running on worker event loop
